更改

+++ JMCLCode/src/main/java/utils/Download.java
@@ -1,7 +1,7 @@
-package Utils;
+package utils;
 
-import Other.PublicVariable;
 import jsonAnalysis.setup.Setup;
+import other.PublicVariable;
 
 import java.io.*;
 import java.net.HttpURLConnection;
@@ -11,15 +11,13 @@ import java.nio.channels.Channels;
 import java.nio.channels.FileChannel;
 import java.nio.channels.ReadableByteChannel;
 
-import static Utils.Utils.getPartOfTheFileContent;
+import static utils.Utils.getPartOfTheFileContent;
 
 public class Download {
-    public static void downloadAFile(URL url, String incompletePath, String name) throws IOException {
+    public static boolean downloadAFile(URL url, String incompletePath, String name) throws IOException {
         File file = new File(incompletePath);
-        if (!file.exists() && !file.isDirectory()) {
-            file.mkdirs();
-        }
-        if (!file.canWrite()) file.setWritable(true);
+        if (!file.exists() && !file.isDirectory()) if (!file.mkdir()) return false;
+        if (!file.canWrite()) if (!file.setWritable(true)) return false;
         //ReadableByteChannel readableByteChannel = Channels.newChannel(url.openStream());
         HttpURLConnection httpURLConnection = (HttpURLConnection) url.openConnection();
         httpURLConnection.setRequestMethod("GET");
@@ -36,14 +34,13 @@ public class Download {
             fileChannel.transferFrom(readableByteChannel, 0, Long.MAX_VALUE);
             fileChannel.close();
         }
-        if (file.length() == 0) {
-            file.delete();
-        }
+        if (file.length() == 0) return file.delete();
+        return true;
     }
 
-    public static void MultiThreadedDownloadAFile(URL url, String path) throws IOException {
+    public static boolean MultiThreadedDownloadAFile(URL url, String path) throws IOException {
         File file = new File(path);
-        if (file.exists() && !file.isDirectory()) file.delete();
+        if (file.exists() && !file.isDirectory()) if (file.delete()) return false;
         URLConnection urlConnection = url.openConnection();
         HttpURLConnection httpURLConnection = (HttpURLConnection) urlConnection;
         httpURLConnection.setRequestMethod("GET");
@@ -55,13 +52,13 @@ public class Download {
         int responseCode = httpURLConnection.getResponseCode();
         if (responseCode == 200) {
             int totalSize = urlConnection.getContentLength();
-            long numberOfSegments = (long) Math.ceil(totalSize / Setup.getSetupInstance().download.threads.multiThreadedDownload.multiThreadedDownloadAFileSegmentSize);
+            long numberOfSegments = (long) Math.ceil((double) totalSize / Setup.getSetupInstance().download.threads.multiThreadedDownload.multiThreadedDownloadAFileSegmentSize);
             for (int i = 0; i <= numberOfSegments; i++) {
                 int finalI = i;
                 PublicVariable.multiThreadedDownloadExecutorService.execute(() -> {
                     try {
-                        if (!file.getParentFile().exists()) file.getParentFile().mkdirs();
-                        if (!file.exists() && !file.isDirectory()) file.createNewFile();
+                        if (!file.getParentFile().exists()) if (file.getParentFile().mkdirs()) return;
+                        if (!file.exists() && !file.isDirectory()) if (file.createNewFile()) return;
                         RandomAccessFile randomAccessFile = new RandomAccessFile(path, "rwd");
                         randomAccessFile.setLength(totalSize);
                         randomAccessFile.seek((long) Setup.getSetupInstance().download.threads.multiThreadedDownload.multiThreadedDownloadAFileSegmentSize * finalI);
@@ -79,9 +76,8 @@ public class Download {
                                 assert bytes != null;
                                 randomAccessFile.write(bytes);
                                 randomAccessFile.close();
-                                if (file.length() == 0) {
-                                    file.delete();
-                                }
+                                if (file.length() == 0)
+                                    if (file.delete()) return;
                                 break;
                             } catch (Throwable ignored) {
                             }
@@ -92,8 +88,7 @@ public class Download {
                 });
             }
         }
-        if (file.length() == 0) {
-            file.delete();
-        }
+        if (file.length() == 0) return !file.delete();
+        return true;
     }
 }
